Add user api to pull json data for individual users
This commit adds two new urls:
'/user'
'/api/user'
The former is a simple form to fill to get information on the individual user.
The latter, if POST with a user, gives JSON information on the individual user listed.
Otherwise if a GET request, it gives a JSON list of all users in the config.
NOTE: This list of users is not inclusive of all users who have filled out the form.
It is only a list of all members who are explicitly listed in the configuration.